Transactions register done

+++ src/main/java/com/tata/test/dto/SaveAccountDto.java
@@ -3,6 +3,8 @@ package com.tata.test.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class SaveAccountDto {
@@ -10,6 +12,6 @@ public class SaveAccountDto {
     private Long clientId;
     private String accountNumber;
     private String accountType;
-    private Double initialBalance;
+    private BigDecimal initialBalance;
 
 }
